+++ example/jest.config.js
@@ -1,0 +1,13 @@
+/* global module */
+module.exports = {
+  collectCoverageFrom: [
+    'src/**/*.js',
+    '!src/**/prop-types.js',
+    '!src/__tests__/**/*.js',
+    '!src/__mocks__/**/*.js',
+  ],
+  setupFiles: ['./jest-setup.js'],
+  setupTestFrameworkScriptFile: './node_modules/jest-enzyme/lib/index.js',
+  snapshotSerializers: ['enzyme-to-json/serializer'],
+  testPathIgnorePatterns: ['<rootDir>/(dist|node_modules)/'],
+}