+++ example/src/components/text.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const Text = () => <div>Hello World</div>
+
+export { Text }